requirements.txt
aiohttp>=3.9,<4
zeroconf>=0.131,<1
PyYAML>=6.0,<7
pyperclip>=1.8,<2
pyautogui>=0.9.54,<1
